move contexts and hooks for user into their own files
