+++ apps/frontend/src/components/Auth/AuthRoute.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from "react";
 import { Navigate } from "react-router-dom";
-import { useUser } from "../../UserContext";
+import useUser from "../../hooks/useUser";
 
 const AuthRoute = ({ children }: { children: ReactNode }) => {
   const { user, loading } = useUser();
