+++ apps/frontend/src/components/Auth/PrivateRoute.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from "react";
 import { Navigate } from "react-router-dom";
-import { useUser } from "../../UserContext";
+import useUser from "../../hooks/useUser";
 
 const PrivateRoute = ({ children }: { children: ReactNode }) => {
   const { user, loading } = useUser();
