+++ apps/frontend/src/contexts/userContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from 'react';
+import { UserContextType } from '../types';
+
+const UserContext = createContext<UserContextType | undefined>(undefined);
+
+export default UserContext;
